Splits the VideoPane even further into MediaBar and TimeLoggerPane.
